% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HES_step2_clean.R
\name{clean_HES}
\alias{clean_HES}
\title{Import HES data}
\usage{
clean_HES(data)
}
\arguments{
\item{data}{a vector containing all the HES years you want to import.}
}
\description{
This function allows you to clean HES data.
}
\keyword{HES}
