% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIH_step2_clean.R
\name{clean_SIH}
\alias{clean_SIH}
\title{Import SIH data}
\usage{
clean_SIH(data)
}
\arguments{
\item{data}{a vector containing all the SIH years you want to import.}
}
\description{
This function allows you to clean SIH data.
}
\keyword{SIH}
