% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_data_dictionary.R
\name{create_data_dictionary}
\alias{create_data_dictionary}
\title{Create html data dictionary}
\usage{
create_data_dictionary(
  data_dictionary_url,
  variable_dictionary,
  years_HH,
  grattandata,
  file_names,
  survey,
  file,
  variable_dictionary_with_new_names,
  ...
)
}
\description{
Create html data dictionary
}
\keyword{internal}
