% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_refyear_list.R
\name{create_refyear_list}
\alias{create_refyear_list}
\title{Create refyear list}
\usage{
create_refyear_list(
  refyears = NULL,
  file_names,
  variable_dictionary,
  survey,
  file,
  ...
)
}
\description{
Figure out what years we want to analyse. If refyears is null it will try and import every year
Otherwise it imports the list provided by the data dictionary
}
\keyword{SIH}
