% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{fix_old_vars}
\alias{fix_old_vars}
\title{Fix old vars}
\usage{
fix_old_vars(data)
}
\arguments{
\item{data}{a tibble created by read_abs_microdata()}
}
\description{
Scale old versions of the HES and the SIH with outdated variables.
}
\keyword{HES}
