% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dictionary.R
\name{import_dictionary}
\alias{import_dictionary}
\title{Import data dictionary}
\usage{
import_dictionary(
  survey,
  file,
  dictionary_file = NULL,
  additional_variables = NULL,
  variable_dictionary_v = 1
)
}
\arguments{
\item{survey}{A string containing the survey you want to import. Usually "hes" or "sih"}

\item{file}{A string containing the survey file of interest - usually "household", "income unit", or "person".}

\item{dictionary_file}{Only to be used if you do not want to use the pre-loaded dictionary. A tibble in the prescribed format in the Vignette.}

\item{additional_variables}{A tibble with at least two columns "r_var_name" and the year of interest. Adds additional variables of interest on top of those that are already in the data dictionary. See vignette for more details}
}
\description{
Import SAS microdata from various ABS datasets, and combine them into a single file with variables given more memorable names.
}
