% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HES_compile.R
\name{import_HES}
\alias{import_HES}
\title{Import HES data}
\usage{
import_HES(data_directory = getwd())
}
\arguments{
\item{data_directory}{A directory with all the required files to run this function. /data contains all the HES DTA files. /HESvariables contains two files: HESkeyvariablenames.csv which is a list of the variables you want in your dataset and filenames.csv contains the filenames of each DTA file in your data folder.}
}
\description{
This function allows you to import HES microdata from various HES years, and combine them into a single tidy dataset.
}
\keyword{HES}
