% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_year.R
\name{\%>\%}
\alias{\%>\%}
\alias{import_year}
\title{Import a single year of your chosen dataset}
\usage{
import_year(filter_cols = TRUE, nrow = Inf, file_names = file_names,
  all_variables = all_variables, year = year,
  grattandata = grattandata, ...)
}
\description{
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\keyword{internal}
