% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HES_helper_fns.R
\name{inflate_HES}
\alias{inflate_HES}
\title{Inflate HES data}
\usage{
inflate_HES(data, to_year)
}
\arguments{
\item{data_directory}{A directory with all the required files to run this function.  /HESvariables contains two files: HESkeyvariablenames.csv which is a list of the variables you want in your dataset and filenames.csv contains the filenames of each DTA file in your data folder.}
}
\description{
This function allows you to inflate HES data. It will only inflate variables listed as 'inflate' in the filename
}
\keyword{HES}
