% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIH_step1_compile.R
\name{read_abs_microdata}
\alias{read_abs_microdata}
\title{Import SIH data}
\usage{
read_abs_microdata(survey = "sih", file = "household",
  refyears = NULL, all_of_latest_year = TRUE,
  additional_variables = NULL, variable_dictionary_v = "1",
  grattandata = FALSE, file_names = NULL, variable_dictionary = NULL)
}
\arguments{
\item{survey}{A string containing the survey you want to import}

\item{file}{A string containing the survey file of interest - usually household, income unit, or person.}

\item{refyears}{A list of numbers containing the years you want to import. If blank all years imported.}

\item{all_of_latest_year}{If TRUE, will import every variable in the most recent dataset. It will also then look at previous years datasets and import variables that have the same name and similar variable labels as the variables in the most recent dataset}

\item{additional_variables}{A tibble with at least two columns "r_var_name" and the year of interest. Adds additional variables of interest on top of those that are already in the data dictionary. See vignette for more details}

\item{variable_dictionary_v}{Version of the preloaded dictionary you wish to use}

\item{grattandata}{Mark true if you want to use the SAS files in the Grattan Institute's data warehouse.}

\item{file_names}{Only to be used if grattandata is false. A tibble with three columns: "Year","Filename" and "Formats", and a row for each year of the dataset you want to import.}

\item{variable_dictionary}{Only to be used if you do not want to use the pre-loaded dictionary. A tibble in the prescribed format in the Vignette.}
}
\description{
Import SAS microdata from various ABS datasets, and combine them into a single file with variables given more memorable names.
}
\keyword{SIH}
