% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIH_step2_clean.R
\name{SIH_unique_hh}
\alias{SIH_unique_hh}
\title{Import SIH data}
\usage{
SIH_unique_hh(data)
}
\arguments{
\item{data}{a vector containing all the SIH years you want to import.}
}
\description{
This function allows you to clean SIH data.
}
\keyword{SIH}
