% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_microdata.R
\name{read_abs_microdata}
\alias{read_abs_microdata}
\title{Import SIH data}
\usage{
read_abs_microdata(survey = "sih", file = "household",
  refyears = NULL, data_dir = getwd(), import_vars = "all",
  additional_variables = NULL, variable_dictionary_v = "1",
  grattandata = FALSE, file_name_file = NULL, dictionary_file = NULL,
  create_html_dictionary = TRUE)
}
\arguments{
\item{survey}{A string containing the survey you want to import. Usually "hes" or "sih"}

\item{file}{A string containing the survey file of interest - usually "household", "income unit", or "person".}

\item{refyears}{A vector containing the years you want to import. If blank all years in the dictionary file are imported.}

\item{data_dir}{a string containing the file path to a directory that contains SAS files for your dataset. Within each subfolder of the directory there must also be a formats file.}

\item{additional_variables}{A tibble with at least two columns "r_var_name" and the year of interest. Adds additional variables of interest on top of those that are already in the data dictionary. See vignette for more details}

\item{variable_dictionary_v}{Version of the preloaded dictionary you wish to use}

\item{grattandata}{Mark true if you want to use the SAS files in the Grattan Institute's data warehouse.}

\item{dictionary_file}{Only to be used if you do not want to use the pre-loaded dictionary. A tibble in the prescribed format in the Vignette.}

\item{create_html_dictionary}{Mark TRUE if you want to create a HTML data dictionary to inspect the labels for each variable.}

\item{file_names}{Only to be used if grattandata is false. A tibble in the prescribed format in the Vignette.}
}
\description{
Import SAS microdata from various ABS datasets, and combine them into a single file with variables given more memorable names.
}
\keyword{SIH}
